app/requirements.txt
kivymd==1.2.0
kivy[base]==2.3.1
m2r2
numpy
onnxruntime
tokenizers
python-docx
pypdf